Update startupJavascript.md
Hi Professor Jensen, I could be totally off-base here, but the above instruction on line 21 says to use deployFiles.sh. This is the command I used to deploy my start up. In these code boxes on line 24 and 30 it says deployService instead of deployFiles. I wasn't sure if this was an alternative way to accomplish the same thing, but deployService didn't work for me. Anyways, I hope this is the right way to do this! Thanks.